ConstantVelocityPositionDateTimeProvider: New GPS source
Introduce a new ConstantVelocityPositionDateTimeProvider source. The source simulates a
10Hz GPS source so every 100ms new Position information is present. The source uses a
list GPS position list to create the new positions. If the position has not enough GPS
positions, new positions are interpolated(Thanks to Lutz ;-)). The interpolation is based
on a constant velocity which is property of the source and can be adjusted.